Better WvStream events, initial tests, WvBuf changes, WvStreamStream, etc.
WvStream onreadable and onwritable events can now have multiple listeners.
WvStream onclose event added.
Moved IWvEventer instance to a static in WvStream, not WvSockStream.
Make IWvStream/WvDbi IDisposable and assert if you forget to dispose.

Add WvStreamClone, WvInBufStream, WvStreamStream, and replace WvFile
(all untested, probably broken).

+++ wvlog.cs
@@ -31,7 +31,7 @@ namespace Wv
 		open_header = null;
 	    }
 	    
-	    uint i;
+	    int i;
 	    while ((i = outbuf.strchr('\n')) > 0)
 	    {
 		if (open_header == null)
@@ -112,7 +112,7 @@ namespace Wv
 		return len; // pretend it's written
 	    
 	    WvBuf outbuf = new WvBuf();
-	    outbuf.put(buf, (uint)offset, (uint)len);
+	    outbuf.put(buf, offset, len);
 	    recv.writelines(wv.fmt("{0}<{1}>: ", name, levelstr(level)),
 			    outbuf);
 	    return len;
